+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 3.0 branch
 ### 3.0.0-RC1
 * requires at least CakePHP 5.0 and PHPUnit 10;
+* fixed `AppTable::findActive()` method, so that it uses the named arguments;
 * added all possible typehints;
 * deleted `BBCode` class;
 * deleted `CreateDirectoriesCommand`, `CreateRobotsCommand`, `CreateVendorsLinksCommand`, `RunAllCommand` and